Add support on Windows to install atexit handler(s) early in console application’s main method.
When running under Visual Studio debugger the console window will not be kept
open when the debugee process terminates. Normally this is not a problem, but there are scenarios,
especially when having errors in configuration, when you would like the console to be kept open so
you can read the output in order to detect the error.

Normally there is an option to start the console application from within Visual Studio but
not attaching the debugger (CTRL+F5) and by default Visual Studio will then inject a pause
in the execution of the process holding up the console, but for some reasons, having more
complex solutions like mono, this feature stops working, so it is not possible to keep the console
open in that scenario for more complex solutions.

This makes it a little problematic to hold the process when working within the debugger in order
to look for console output and errors. Since mono can exit the process using exit command and not
only by returning from main, the developer needs to set breakpoints on both end of main and in c-runtime
exit methods. Even if that is a possible workaround this option won't work when running the process outside
the debugger (CTRL+F5).

This commit adds a command line argument that can be used on Windows (potential on other platforms as well), --atexit.
It will install a number of recognized exit handlers at startup. Currently only one is defined, waitkeypress, but it can
be extended with more, if needed. The waitkeypress will hold the console and process open until a key is pressed and is very
useful when running and debugging mono from within Visual Studio. Example of output when running mono’s mono-mini-regression-test
with an assembly not found together with –atexit=waitkeypress command line argument:

failed to load assembly: gshared-nofound.exe
Overall results: tests: 0, 100% pass, opt combinations: 25
Press any key to continue . . .

Currently this is only enabled for Windows, but could be enabled for other platforms,
if similar need exist elsewhere.

There is also an option to always install the default atexit handler through a define
when building mono:

MONO_INSTALL_DEFAULT_ATEXIT_COMMANDS

This can be a convenient way to always include the waitkeypress when
actively working within Visual Studio and running debugging targets without the need to change
any command line configuration of the target test projects.