Reduce warnings in mono Visual Studio project.
Currently a lot of warnings are produced when building the mono solution on Windows
and due to this, some projects are built on a very low warning level while others produce
too much warnings potentially cloaking real problems/errors. The ambition should be to build
all mono projects on warning level 4 with 0 outstanding warnings and handle all generated
warnings in correct way (either by adjusting source code when possible or if that not possible
solve warning use alternative techniques). This is needed in order to spot new
warnings as they arise since they are a valuable aid during development.

This commit is the first commit that will increase warning level and reduce the warnings in
one Visual Studio project, mono.vcxproj generating mono(-sgen).exe. Previously this
projects were build using warning level 1!!!, but this commit bumps it to warning level 4 and
makes sure the project (simple starting case since it just includes one source, main.c) builds
without warnings.

In order to reduce the warnings in mono.vcxproj, two different strategies where applied. First, for warnings
that are related to use of nonstandard C extensions supported on all other platforms, the warning is disabled
for the complete project (in this case 4214 and 4204). For other warnings that are mostly related to down casts
of scalar data types, like UINT64 -> UINT, a set of helper macros have been added in mono-typecast-helper.h in
libmonoutils. These macros can be used to do the needed cast to explicitly show the intent of a downcast,
if that’s what’s intended and not a bug detected and correctly reported by the compiler (that needs to be fixed).

There are several benefits of doing the explicit downcasts using the helper macros, compare to plain c casts. First it clearly
shows the intent of the downcast and will be easier to locate in the code base. Secondly, it is a way to resolve the generated warnings
in a controlled and trackable way, if the intent is actually a downcast, since other solutions might not be feasible.
Third, there is an implementation of the macros that includes diagnostic and error detection logic enabled under debug builds
(but could be adjusted to be included in release builds as well using a different diagnostic method). This diagnostic will validate
that the downcast is correctly done during runtime on live data, validating the assumption of the downcast when running for example
test suites. Fourth, possible to change how the underlying cast is performed overtime.